+++ src/app/shared/components/table/table.component.ts
@@ -19,7 +19,7 @@ export class TableComponent<T> implements OnInit {
   }
 
   private initColumns(): void {
-    this.columns = this.dataSource.length ? Object.keys(this.dataSource[0]) : [];
+    this.columns = this.dataSource.length ? Object.keys(this.dataSource[0]) : ['No data'];
   }
 
   public onRowClick(row: T): void {