+++ src/app/views/matches/matches.module.ts
@@ -1,10 +1,12 @@
 import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
+import { SharedModule } from '@shared/shared.module';
 import { MatchesRoutingModule } from './matches-routing.module';
 import { MatchesComponent } from './matches.component';
 
 @NgModule({
-  imports: [MatchesRoutingModule],
+  imports: [CommonModule, SharedModule, MatchesRoutingModule],
   exports: [],
   declarations: [MatchesComponent],
   providers: []